Create Loading component and add userReucer to store
