+++ package.json
@@ -14,6 +14,7 @@
     "react": "^17.0.2",
     "react-bootstrap": "^1.6.1",
     "react-dom": "^17.0.2",
+    "react-fontawesome": "^1.7.1",
     "react-hook-form": "^7.11.1",
     "react-loader-spinner": "^4.0.0",
     "react-redux": "^7.2.4",
