+++ src/containers/Loading.js
@@ -1,0 +1,6 @@
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { faSpinner } from '@fortawesome/free-solid-svg-icons';
+
+const Loading = () => (<div className=""><FontAwesomeIcon icon={faSpinner} spin size="10x" /></div>);
+
+export default Loading;
